Restructuring relative paths for MVT
